Add MasmItemsPredictor
VS ships with masm.props / masm.targets:
- $(VCTargetsPath)\BuildCustomizations\masm.props
- $(VCTargetsPath)\BuildCustomizations\masm.targets

They are used to compile .asm files via the ml64.exe assembler.

The targets the input .asm files @(MASM), as well as a set of include
directories that the assembler pre-processor will use to search
for files included in the .asm file %(MASM.IncludePaths).

The targets will write the output object file to %(MASM.ObjectFileName).
It also supports a few other options which will emit files, which the
predictor also supports (AssembledCodeListingFile, BrowseFile).